Refactor
Renamed several files :
TaskBag.java -> Bag.java
TaskBagProcessor.java -> BagProcessor.java
TaskQueue.java -> TaskBag.java
Queue -> TaskQueue.java

Also checked the Javadoc and corrected some unconsistencies
