+++ apgas.glb/src/apgas/glb/util/Task.java
@@ -3,10 +3,8 @@
  */
 package apgas.glb.util;
 
-import apgas.glb.TaskBagProcessor;
-
 /**
- * Abstract Task to be be run.
+ * Abstract Task to be be run and handled by a {@link TaskBag}
  *
  * @author Patrick Finnerty
  *
@@ -16,20 +14,19 @@ public interface Task {
    * Processes this task.
    * <p>
    * Can (possibly) spawn other tasks and send them in the queue by calling the
-   * available methods of the {@link TaskBagProcessor} provided by
-   * {@link #setProcessor(TaskBagProcessor)}
+   * available methods of the {@link TaskBag} provided by
+   * {@link #setProcessor(TaskBag)}.
    */
   public void process();
 
   /**
-   * Set the {@link TaskBagProcessor} responsible for processing this task. If
-   * the task needs to spawn new tasks, it can then add them to this
-   * {@link TaskBagProcessor} which should be kept as a class member. However if
-   * that is not the case, the implementation can simply leave this method
-   * empty.
+   * Set the {@link TaskBag} responsible for processing this task. If the task
+   * needs to spawn new tasks, it should keep the provided {@link TaskBag} as a
+   * class member. However if that is not the case, the implementation can
+   * simply leave this method empty.
    *
    * @param p
-   *          the {@link TaskBagProcessor}
+   *          the {@link TaskBag} responsible for handling the {@link Task}
    */
-  public void setProcessor(TaskQueue p);
+  public void setProcessor(TaskBag p);
 }
