+++ apgas.glb/src/apgas/glb/util/package-info.java
@@ -2,8 +2,8 @@
  *
  */
 /**
- * Package apgas.glb.util contains a class Queue which implements
- * {@link apgas.glb.TaskBag} and can be used by programmers to handle
+ * Package apgas.glb.util contains a class TaskQueue which implements
+ * {@link apgas.glb.Bag} and can be used by programmers to handle
  * {@link apgas.glb.util.Task}s that are independant from one another.
  *
  * @author Patrick Finnerty
